+++ DragNDrop/Base.lproj/Main.storyboard
@@ -5,10 +5,10 @@
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="9529"/>
     </dependencies>
     <scenes>
-        <!--View Controller-->
+        <!--DragN Drop View Controller-->
         <scene sceneID="h9y-kV-6G7">
             <objects>
-                <tableViewController storyboardIdentifier="ViewController" id="6LF-u0-T5v" customClass="ViewController" customModule="DragNDrop" customModuleProvider="target" sceneMemberID="viewController">
+                <tableViewController storyboardIdentifier="DragNDropViewController" id="6LF-u0-T5v" customClass="DragNDropViewController" customModule="DragNDrop" customModuleProvider="target" sceneMemberID="viewController">
                     <tableView key="view" clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="44" sectionHeaderHeight="28" sectionFooterHeight="28" id="b9k-bb-yUh">
                         <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
