+++ DragNDrop/DragNDropViewController.swift
@@ -9,89 +9,167 @@
 import UIKit
 
 class DragNDropViewController: UITableViewController {
-
+    
+    var itemsArray : [String]
+    
+    required init(coder aDecoder: NSCoder) {
+        itemsArray = [String]()
+        
+        let item1 = "Bananas"
+        let item2 = "Oranges"
+        let item3 = "Kale"
+        let item4 = "Milk"
+        let item5 = "Yogurt"
+        let item6 = "Crackers"
+        let item7 = "Cheese"
+        let item8 = "Carrots"
+        let item9 = "Ice Cream"
+        let item10 = "Olive Oil"
+        
+        itemsArray.append(item1)
+        itemsArray.append(item2)
+        itemsArray.append(item3)
+        itemsArray.append(item4)
+        itemsArray.append(item5)
+        itemsArray.append(item6)
+        itemsArray.append(item7)
+        itemsArray.append(item8)
+        itemsArray.append(item9)
+        itemsArray.append(item10)
+        
+        super.init(coder: aDecoder)!
+    }
+    
     override func viewDidLoad() {
         super.viewDidLoad()
-
-        // Uncomment the following line to preserve selection between presentations
-        // self.clearsSelectionOnViewWillAppear = false
-
-        // Uncomment the following line to display an Edit button in the navigation bar for this view controller.
-        // self.navigationItem.rightBarButtonItem = self.editButtonItem()
+        
+        tableView.contentInset = UIEdgeInsets(top: 20, left: 0, bottom: 0, right: 0)
+        
+        let longpress = UILongPressGestureRecognizer(target: self, action: "longPressGestureRecognized:")
+        tableView.addGestureRecognizer(longpress)
     }
-
+    
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
     }
-
+    
+    func longPressGestureRecognized(gestureRecognizer: UIGestureRecognizer) {
+        let longPress = gestureRecognizer as! UILongPressGestureRecognizer
+        let state = longPress.state
+        let locationInView = longPress.locationInView(tableView)
+        let indexPath = tableView.indexPathForRowAtPoint(locationInView)
+        
+        struct My {
+            static var cellSnapshot : UIView? = nil
+            static var cellIsAnimating : Bool = false
+            static var cellNeedToShow : Bool = false
+        }
+        struct Path {
+            static var initialIndexPath : NSIndexPath? = nil
+        }
+        
+        switch state {
+        case UIGestureRecognizerState.Began:
+            if indexPath != nil {
+                Path.initialIndexPath = indexPath
+                let cell = tableView.cellForRowAtIndexPath(indexPath!) as UITableViewCell!
+                My.cellSnapshot  = snapshotOfCell(cell)
+                
+                var center = cell.center
+                My.cellSnapshot!.center = center
+                My.cellSnapshot!.alpha = 0.0
+                tableView.addSubview(My.cellSnapshot!)
+                
+                UIView.animateWithDuration(0.25, animations: { () -> Void in
+                    center.y = locationInView.y
+                    My.cellIsAnimating = true
+                    My.cellSnapshot!.center = center
+                    My.cellSnapshot!.transform = CGAffineTransformMakeScale(1.05, 1.05)
+                    My.cellSnapshot!.alpha = 0.98
+                    cell.alpha = 0.0
+                    }, completion: { (finished) -> Void in
+                        if finished {
+                            My.cellIsAnimating = false
+                            if My.cellNeedToShow {
+                                My.cellNeedToShow = false
+                                UIView.animateWithDuration(0.25, animations: { () -> Void in
+                                    cell.alpha = 1
+                                })
+                            } else {
+                                cell.hidden = true
+                            }
+                        }
+                })
+            }
+            
+        case UIGestureRecognizerState.Changed:
+            if My.cellSnapshot != nil {
+                var center = My.cellSnapshot!.center
+                center.y = locationInView.y
+                My.cellSnapshot!.center = center
+                
+                if ((indexPath != nil) && (indexPath != Path.initialIndexPath)) {
+                    itemsArray.insert(itemsArray.removeAtIndex(Path.initialIndexPath!.row), atIndex: indexPath!.row)
+                    tableView.moveRowAtIndexPath(Path.initialIndexPath!, toIndexPath: indexPath!)
+                    Path.initialIndexPath = indexPath
+                }
+            }
+        default:
+            if Path.initialIndexPath != nil {
+                let cell = tableView.cellForRowAtIndexPath(Path.initialIndexPath!) as UITableViewCell!
+                if My.cellIsAnimating {
+                    My.cellNeedToShow = true
+                } else {
+                    cell.hidden = false
+                    cell.alpha = 0.0
+                }
+                
+                UIView.animateWithDuration(0.25, animations: { () -> Void in
+                    My.cellSnapshot!.center = cell.center
+                    My.cellSnapshot!.transform = CGAffineTransformIdentity
+                    My.cellSnapshot!.alpha = 0.0
+                    cell.alpha = 1.0
+                    
+                    }, completion: { (finished) -> Void in
+                        if finished {
+                            Path.initialIndexPath = nil
+                            My.cellSnapshot!.removeFromSuperview()
+                            My.cellSnapshot = nil
+                        }
+                })
+            }
+        }
+    }
+    
+    func snapshotOfCell(inputView: UIView) -> UIView {
+        UIGraphicsBeginImageContextWithOptions(inputView.bounds.size, false, 0.0)
+        inputView.layer.renderInContext(UIGraphicsGetCurrentContext()!)
+        let image = UIGraphicsGetImageFromCurrentImageContext() as UIImage
+        UIGraphicsEndImageContext()
+        
+        let cellSnapshot : UIView = UIImageView(image: image)
+        cellSnapshot.layer.masksToBounds = false
+        cellSnapshot.layer.cornerRadius = 0.0
+        cellSnapshot.layer.shadowOffset = CGSizeMake(-5.0, 0.0)
+        cellSnapshot.layer.shadowRadius = 5.0
+        cellSnapshot.layer.shadowOpacity = 0.4
+        return cellSnapshot
+    }
+    
     // MARK: - Table view data source
-
-    override func numberOfSectionsInTableView(tableView: UITableView) -> Int {
-        // #warning Potentially incomplete method implementation.
-        // Return the number of sections.
-        return 0
-    }
-
+    
     override func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        // #warning Incomplete method implementation.
-        // Return the number of rows in the section.
-        return 0
+        return itemsArray.count
     }
-
-    /*
+    
     override func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCellWithIdentifier("reuseIdentifier", forIndexPath: indexPath) as UITableViewCell
-
-        // Configure the cell...
-
+        let cell = tableView.dequeueReusableCellWithIdentifier("Cell", forIndexPath: indexPath) as UITableViewCell
+        cell.textLabel?.text = itemsArray[indexPath.row]
         return cell
     }
-    */
-
-    /*
-    // Override to support conditional editing of the table view.
-    override func tableView(tableView: UITableView, canEditRowAtIndexPath indexPath: NSIndexPath) -> Bool {
-        // Return NO if you do not want the specified item to be editable.
-        return true
+    
+    override func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
+        tableView.deselectRowAtIndexPath(indexPath, animated: false)
     }
-    */
-
-    /*
-    // Override to support editing the table view.
-    override func tableView(tableView: UITableView, commitEditingStyle editingStyle: UITableViewCellEditingStyle, forRowAtIndexPath indexPath: NSIndexPath) {
-        if editingStyle == .Delete {
-            // Delete the row from the data source
-            tableView.deleteRowsAtIndexPaths([indexPath], withRowAnimation: .Fade)
-        } else if editingStyle == .Insert {
-            // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
-        }    
-    }
-    */
-
-    /*
-    // Override to support rearranging the table view.
-    override func tableView(tableView: UITableView, moveRowAtIndexPath fromIndexPath: NSIndexPath, toIndexPath: NSIndexPath) {
-
-    }
-    */
-
-    /*
-    // Override to support conditional rearranging of the table view.
-    override func tableView(tableView: UITableView, canMoveRowAtIndexPath indexPath: NSIndexPath) -> Bool {
-        // Return NO if you do not want the item to be re-orderable.
-        return true
-    }
-    */
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
-        // Get the new view controller using [segue destinationViewController].
-        // Pass the selected object to the new view controller.
-    }
-    */
-
 }
